+++ app/scripts/TrackRenderer.js
@@ -101,6 +101,7 @@ class TrackRenderer extends React.Component {
 
     this.yPositionOffset = 0;
     this.xPositionOffset = 0;
+    this.scrollTop = 0;
 
     this.scrollTimeout = null;
     this.activeTransitions = 0;
@@ -204,6 +205,7 @@ class TrackRenderer extends React.Component {
     this.boundForwardEvent = this.forwardEvent.bind(this);
     this.boundScrollEvent = this.scrollEvent.bind(this);
     this.boundForwardContextMenu = this.forwardContextMenu.bind(this);
+    this.onScrollHandlerBound = this.onScrollHandler.bind(this);
   }
 
   // eslint-disable-next-line camelcase
@@ -371,6 +373,14 @@ class TrackRenderer extends React.Component {
       }
     }
 
+    if (prevProps.zoomable !== this.props.zoomable) {
+      if (this.props.zoomable) {
+        this.addZoom();
+      } else {
+        this.removeZoom();
+      }
+    }
+
     this.addEventTracker();
   }
 
@@ -456,7 +466,7 @@ class TrackRenderer extends React.Component {
     this.pMask.beginFill();
     this.pMask.drawRect(
       this.xPositionOffset,
-      this.yPositionOffset,
+      this.yPositionOffset + this.scrollTop,
       this.currentProps.width,
       this.currentProps.height
     );
@@ -1874,6 +1884,12 @@ class TrackRenderer extends React.Component {
     this.props.pubSub.publish('app.event', e);
   }
 
+  onScrollHandler(e) {
+    this.scrollTop = e.target.scrollTop;
+    this.setMask();
+    this.props.pubSub.publish('app.viewScroll', this.scrollTop);
+  }
+
   /* ------------------------------- Render ------------------------------- */
 
   render() {
@@ -1895,6 +1911,11 @@ class TrackRenderer extends React.Component {
         <div
           ref={(c) => { this.eventTracker = c; }}
           className="track-renderer-events"
+          onScroll={this.onScrollHandlerBound}
+          style={{
+            overflowX: 'hidden',
+            overflowY: this.props.scrollable ? 'auto' : 'hidden',
+          }}
           styleName="track-renderer-events"
         >
           {this.currentProps.children}
@@ -1921,6 +1942,7 @@ TrackRenderer.defaultProps = {
   paddingLeft: 0,
   paddingTop: 0,
   positionedTracks: [],
+  scrollable: false,
   topHeight: 0,
   topHeightNoGallery: 0,
   width: 0,
@@ -1950,6 +1972,7 @@ TrackRenderer.propTypes = {
   pixiStage: PropTypes.object.isRequired,
   pluginTracks: PropTypes.object,
   positionedTracks: PropTypes.array,
+  scrollable: PropTypes.bool,
   setCentersFunction: PropTypes.func,
   svgElement: PropTypes.object.isRequired,
   theme: PropTypes.symbol.isRequired,
@@ -1960,6 +1983,7 @@ TrackRenderer.propTypes = {
   xDomainLimits: PropTypes.array,
   yDomainLimits: PropTypes.array,
   valueScaleZoom: PropTypes.bool,
+  zoomable: PropTypes.bool.isRequired,
   zoomDomain: PropTypes.array,
 };